refactor(api): Phase 13 step2 — split mutations/mod.rs into domain modules [GREEN]
Extract custom_mutations content into separate domain modules:
- auth.rs (436L): auth/user mutations + UserOutput/SignUpOutput/SignInOutput
- dog.rs (359L): CRUD dog mutations + BirthDate/DogOutput types
- dog_member.rs (292L): invitation + member mutations + output types
- walk.rs (441L): walk lifecycle mutations + WalkOutput/WalkerOutput
- walk_event.rs (198L): walk event mutation + WalkEventOutput
- encounter.rs (120L): encounter mutations
- photo.rs (124L): presigned URL mutations + PresignedUrlOutput
- mod.rs (47L): submodule declarations, pub use re-exports, mutation_fields()

All files under 500 lines. No logic changes — mechanical code movement only.
cargo test --lib: 61/61 pass. All integration tests: 61/61 pass.
